kv/KeyValueHistogram: make Bluestore's DBHistogram class reusable.
